Timeline working with audio clip.
First version of the timeline is implemented. The video file plays over
the timeline normally. The audio clip plays once at the beginning and
then later on if the user adds it to the timeline.

Things I broke:
Tracking no longer works.
The click to advance to certain positions in the timeline is broken.
Timeline doesn't automatically loop.